chore: move fill_with_equilibrium to src/host/mod.rs and use it
use fill_with_equilibrium() to fill equilibrium in output capture
